Refactoring code to remove unused and add tests
